fix(scheduler): publish resource-usage origin event from ApiWorkerScheduler
The worker reports observed peak memory in ExecuteResult.resource_usage,
which WorkerApiServer forwards via WorkerScheduler::record_action_resource_usage.
That trait method has a no-op default; the only override lived on
SimpleScheduler. But WorkerApiServer is handed the inner ApiWorkerScheduler
(the worker_scheduler returned by SimpleScheduler::new), which did not
override it — so every worker-reported resource usage hit the no-op and
observed_worker_peak_memory_mib was never populated.

Give ApiWorkerScheduler the origin_event_tx and implement
record_action_resource_usage there (it already owns running_action_info),
publishing the ResponseEvent::ActionResourceUsage origin event that bep-etl
ingests. Verified via debug build: worker samples peak, sends it, scheduler
receives it — but record_action_resource_usage's publish was never reached.
