+++ nativelink-scheduler/src/api_worker_scheduler.rs
@@ -26,13 +26,19 @@ use nativelink_metric::{
     MetricFieldData, MetricKind, MetricPublishKnownKindData, MetricsComponent,
     RootMetricsComponent, group,
 };
+use nativelink_proto::com::github::trace_machina::nativelink::events::{
+    Event, OriginEvent, ResponseEvent, event, response_event,
+};
+use nativelink_proto::com::github::trace_machina::nativelink::remote_execution::ActionResourceUsage;
 use nativelink_util::action_messages::{OperationId, WorkerId};
 use nativelink_util::operation_state_manager::{UpdateOperationType, WorkerStateManager};
+use nativelink_util::origin_event::get_node_id;
 use nativelink_util::platform_properties::PlatformProperties;
 use nativelink_util::shutdown_guard::ShutdownGuard;
-use tokio::sync::Notify;
+use tokio::sync::{Notify, mpsc};
 use tonic::async_trait;
 use tracing::{error, info, trace, warn};
+use uuid::Uuid;
 
 /// Metrics for tracking scheduler performance.
 #[derive(Debug, Default)]
@@ -491,6 +497,10 @@ pub struct ApiWorkerScheduler {
 
     /// Performance metrics for observability.
     metrics: Arc<SchedulerMetrics>,
+
+    /// Channel for publishing origin events such as worker-observed action
+    /// resource usage. `None` when origin events are disabled.
+    maybe_origin_event_tx: Option<mpsc::Sender<OriginEvent>>,
 }
 
 impl ApiWorkerScheduler {
@@ -501,6 +511,7 @@ impl ApiWorkerScheduler {
         worker_change_notify: Arc<Notify>,
         worker_timeout_s: u64,
         worker_registry: SharedWorkerRegistry,
+        maybe_origin_event_tx: Option<mpsc::Sender<OriginEvent>>,
     ) -> Arc<Self> {
         Arc::new(Self {
             inner: Mutex::new(ApiWorkerSchedulerImpl {
@@ -516,6 +527,7 @@ impl ApiWorkerScheduler {
             worker_timeout_s,
             worker_registry,
             metrics: Arc::new(SchedulerMetrics::default()),
+            maybe_origin_event_tx,
         })
     }
 
@@ -622,6 +634,56 @@ impl ApiWorkerScheduler {
 impl WorkerScheduler for ApiWorkerScheduler {
     fn get_platform_property_manager(&self) -> &PlatformPropertyManager {
         self.platform_property_manager.as_ref()
+    }
+
+    async fn record_action_resource_usage(
+        &self,
+        worker_id: &WorkerId,
+        operation_id: &OperationId,
+        mut resource_usage: ActionResourceUsage,
+    ) -> Result<(), Error> {
+        // The worker API talks to this `ApiWorkerScheduler` (it is the
+        // `WorkerScheduler` returned by `SimpleScheduler::new`), so the
+        // resource-usage origin event must be published here. Previously the
+        // only override lived on `SimpleScheduler`, which this path never
+        // reaches, so the event was silently dropped by the trait's no-op
+        // default and `observed_worker_peak_memory_mib` was never recorded.
+        let Some(origin_event_tx) = self.maybe_origin_event_tx.as_ref() else {
+            return Ok(());
+        };
+        let Some(action_info) = self.running_action_info(worker_id, operation_id).await else {
+            return Ok(());
+        };
+
+        if resource_usage.operation_id.is_empty() {
+            resource_usage.operation_id = operation_id.to_string();
+        }
+        if resource_usage.worker_id.is_empty() {
+            resource_usage.worker_id = worker_id.to_string();
+        }
+
+        let event = Event {
+            event: Some(event::Event::Response(ResponseEvent {
+                event: Some(response_event::Event::ActionResourceUsage(resource_usage)),
+            })),
+        };
+        let origin_event = OriginEvent {
+            version: 0,
+            event_id: Uuid::now_v6(&get_node_id(Some(&event)))
+                .hyphenated()
+                .to_string(),
+            parent_event_id: action_info
+                .scheduler_start_execute_event_id
+                .clone()
+                .unwrap_or_default(),
+            bazel_request_metadata: action_info.origin_metadata.bazel_metadata.clone(),
+            identity: action_info.origin_metadata.identity,
+            event: Some(event),
+        };
+        if let Err(err) = origin_event_tx.try_send(origin_event) {
+            warn!(?err, "Failed to publish action resource usage origin event");
+        }
+        Ok(())
     }
 
     async fn add_worker(&self, worker: Worker) -> Result<(), Error> {
